requirements.txt
requests==2.*
google-cloud-bigquery==3.*